Added grade field to CourseStudentMapper [UM-14]

+++ src/main/java/com/archisacadeny/student/CourseStudentMapper.java
@@ -21,7 +21,7 @@ public class CourseStudentMapper {
                             id INTEGER DEFAULT nextval('s_c_mapper_id') PRIMARY KEY NOT NULL,
                             course_id INTEGER NOT NULL,
                             student_id INTEGER NOT NULL,
-                            grade INTEGER,
+                            grade FLOAT,
                     CONSTRAINT fk_course_id FOREIGN KEY (course_id) REFERENCES "public"."courses"(id),
                     CONSTRAINT fk_student_id FOREIGN KEY (student_id) REFERENCES "public"."students"(id)
                     );
